Importing schema to migrate
git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@1807 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2

+++ Kistl.API.Server/ISchemaProvider.cs
@@ -54,7 +54,7 @@ namespace Kistl.API.Server
     {
         public string Name { get; set; }
         public System.Data.DbType Type { get; set; }
-        public long Size { get; set; }
+        public int Size { get; set; }
         public bool IsNullable { get; set; }
 
         public override string ToString()
